requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-jose[cryptography]
requests
pytest
httpx

psycopg2-binary
sqlalchemy
